[DEV-3829] Refactoring
